requirements.txt
os
re
uuid
datetime
Flask==3.0.2
azure-core== 1.30.2
azure-search-documents==11.5.1
openai== 1.46.1
azure-cosmos==4.7.0
PyMuPDF==1.24.10
python-dotenv==1.0.1

